requirements.txt
APScheduler==3.11.0
playwright==1.52.0
pyee==13.0.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
greenlet==3.2.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
requests==2.32.3
flask==3.1.0
gunicorn==21.2.0
